# swagger.yaml
openapi: 3.0.0
info:
  title: IMF Gadget API
  version: 1.0.0
  description: API for managing Impossible Missions Force gadgets
servers:
  - url: http://localhost:3000
paths:
  /register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '201':
          description: User registered
        '400':
          description: Missing fields
        '409':
          description: User already exists
  /login:
    post:
      summary: Login and get JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: JWT token
        '400':
          description: Missing fields
        '401':
          description: Invalid credentials
  /gadgets:
    get:
      summary: Get all gadgets
      parameters:
        - in: query
          name: status
          schema:
            type: string
          description: Filter by status
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of gadgets
    post:
      summary: Add a new gadget
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                status:
                  type: string
      responses:
        '201':
          description: Gadget created
  /gadgets/{id}:
    patch:
      summary: Update a gadget
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                status:
                  type: string
      responses:
        '200':
          description: Gadget updated
        '404':
          description: Gadget not found
    delete:
      summary: Decommission a gadget
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Gadget decommissioned
        '404':
          description: Gadget not found
  /gadgets/{id}/self-destruct:
    post:
      summary: Trigger self-destruct sequence
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Self-destruct confirmation code
        '404':
          description: Gadget not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT 